basic test cases done

+++ controller/controller.js
@@ -9,16 +9,17 @@ var model = require('./../models/model');
 	@description function to get data from model show existing tasks on server load
 	@param {server object request , server object response}
 */
-function showTodo(req,res){
+function showTodo(req,res) {
 	model.showTodo(req).then(
 		function(data) {
 			res.render('todo.html' , {
 				welcomeMessage: 'To-Do Checklist App' ,
 				todoFileTxtShow: data
 			});//render close
-		},
-		function(err){if(err)throw err;}
-	)//then close
+		}
+	).catch((err)=>{
+		console.log(err)
+	})//then close
 }
 
 
@@ -30,12 +31,14 @@ function showTodo(req,res){
 */
 function addTodo(req,res) {
 	var task = req.body.data;
-	if(task.data === "") {
-		throw err;
+	console.log("task:",task);
+	if(!task.trim()) {
+		res.status(400).send({error:"Enter a valid input"})
 	}
 	else {
 		model.addTodo(task).then(
 			function(data) {
+				console.log("cnt:",data);
 				res.send(data);
 			},
 			function(err){if(err)throw err;}
@@ -52,12 +55,18 @@ function addTodo(req,res) {
 */
 function deleteTodo(req,res){
 	var taskDestroy = req.body;
-	model.deleteTodo(taskDestroy).then(
-		function(data) {
-      res.send("success");
-		},
-		function(err){if(err)throw err;}
-	)//then close
+	if(taskDestroy.id == "") {
+		console.log("Enter an id");
+	}
+	else {
+		model.deleteTodo(taskDestroy).then(
+			function(data) {
+				console.log("cnt",data)
+	      res.send("success");
+			},
+			function(err){if(err)throw err;}
+		)//then close
+	}
 }
 
 
@@ -91,7 +100,7 @@ function toggleStatus(req,res) {
 		array/file and get response and send it back to client
 	@param {server object request , server object response}
 */
-function toggleAll(req,res){
+function toggleAll(req,res) {
 	var condition = req.query
 	model.toggleAll(condition).then(
 		function(data) {
@@ -108,10 +117,14 @@ function toggleAll(req,res){
 		array/file and get response and send it back to client
 	@param {server object request , server object response}
 */
-function clearCompleted(req,res){
+function clearCompleted(req,res) {
 	model.clearCompleted().then(
 		function(data) {
-      res.send("success");
+			var tmp;
+			for(i=0; i<data.length; i++) {
+				tmp = data[i].activeStatus;
+			}
+      res.send(tmp);
 		},
 		function(err){if(err)throw err;}
 	)//then close
@@ -157,9 +170,10 @@ function getCompleted(req,res){
 function alterTask(req,res){
 	var text = req.body;
 	var taskId = req.params.textId;
+	console.log(text,taskId);
 	model.alterTask(taskId , text).then(
 		function(data) {
-      res.send("success");
+      res.send(data);
 		},
 		function(err){if(err)throw err;}
 	)//then close
